unify install module
