+++ src/pymmcore_plus/_cli.py
@@ -1,24 +1,15 @@
-import os
-import re
 import shutil
-import subprocess
 import sys
-import tempfile
-from contextlib import contextmanager, suppress
+from contextlib import suppress
 from pathlib import Path
-from platform import system
-from typing import Dict, Iterator, Optional
-from urllib.request import urlopen, urlretrieve
+from typing import Optional
 
 import pymmcore_plus
 import typer
 from pymmcore_plus._logger import set_log_level
 from pymmcore_plus._util import USER_DATA_MM_PATH
-from rich import print, progress
+from rich import print
 
-PLATFORM = system()
-BASE_URL = "https://download.micro-manager.org"
-_list = list
 
 app = typer.Typer(no_args_is_help=True, add_completion=False)
 
@@ -71,8 +62,8 @@ def clean() -> None:
         print(":sparkles: [bold green]No files to remove.")
 
 
-@app.command()
-def list() -> None:  # noqa: A001
+@app.command(name="list")
+def _list() -> None:
     """Show all Micro-Manager installs downloaded by pymmcore-plus."""
     if USER_DATA_MM_PATH.exists():
         print(f":file_folder:[bold green] {USER_DATA_MM_PATH}")
@@ -120,159 +111,9 @@ def install(
     ),
 ) -> None:
     """Install Micro-Manager Device adapters."""
-    if PLATFORM not in ("Darwin", "Windows"):  # pragma: no cover
-        print(f":x: [bold red]Unsupported platform: {PLATFORM!r}")
-        raise typer.Exit(1)
+    import pymmcore_plus.install
 
-    if release == "latest":
-        plat = {
-            "Darwin": "macos/Micro-Manager-x86_64-latest.dmg",
-            "Windows": "windows/MMSetup_x64_latest.exe",
-        }[PLATFORM]
-        url = f"{BASE_URL}/latest/{plat}"
-    else:
-        available = _available_versions()
-        if release not in available:
-            n = 15
-            avail = ", ".join(_list(available)[:n]) + " ..."
-            raise typer.BadParameter(
-                f"Release {release!r} not found. Last {n} releases:\n{avail}"
-            )
-        url = available[release]
-
-    with tempfile.TemporaryDirectory() as tmpdir:
-        installer = Path(tmpdir) / url.split("/")[-1]
-        _download_url(url=url, output_path=installer)
-        if PLATFORM == "Darwin":
-            _mac_install(installer, dest)
-        elif PLATFORM == "Windows":
-            # for windows, we need to know the latest version
-            filename = _get_download_name(url)
-            filename = filename.replace("MMSetup_64bit", "Micro-Manager")
-            filename = filename.replace(".exe", "")
-            _win_install(installer, dest / filename)
-
-    print(f":sparkles: [bold green]Installed to {dest}![/bold green] :sparkles:")
-
-
-def _get_download_name(url: str) -> str:
-    """Return the name of the file to be downloaded from `url`."""
-    with urlopen(url) as tmp:
-        content: str = tmp.headers.get("Content-Disposition")
-        for part in content.split(";"):
-            if "filename=" in part:
-                return part.split("=")[1].strip('"')
-    return ""
-
-
-@contextmanager
-def _spinner(
-    text: str = "Processing...", color: str = "bold blue"
-) -> Iterator[progress.Progress]:
-    with progress.Progress(
-        progress.SpinnerColumn(),
-        progress.TextColumn(f"[{color}]{text}"),
-        transient=True,
-    ) as pbar:
-        pbar.add_task(description=text, total=None)
-        yield pbar
-
-
-def _win_install(exe: Path, dest: Path) -> None:
-    cmd = [str(exe), "/VERYSILENT", "/SUPPRESSMSGBOXES", "/NORESTART", f"/DIR={dest}"]
-    with _spinner("Installing ..."):
-        subprocess.run(cmd, check=True)
-
-
-def _mac_install(dmg: Path, dest: Path) -> None:
-    """Install Micro-Manager `dmg` to `dest`."""
-    # with progress bar, mount dmg
-    with _spinner("Mounting ..."):
-        proc = subprocess.run(
-            ["hdiutil", "attach", "-nobrowse", str(dmg)],
-            capture_output=True,
-        )
-        if proc.returncode != 0:  # pragma: no cover
-            typer.secho(
-                f"\nError mounting {dmg.name}:\n{proc.stderr.decode()}",
-                fg="bright_red",
-            )
-            raise typer.Exit(code=proc.returncode)
-
-    # with progress bar, mount dmg
-    with _spinner(f"Installing to {dest} ..."):
-        # get mount point
-        mount = proc.stdout.splitlines()[-1].split()[-1].decode()
-        try:
-            try:
-                src = next(Path(mount).glob("Micro-Manager*"))
-            except StopIteration:  # pragma: no cover
-                typer.secho(
-                    "\nError: Could not find Micro-Manager in dmg.\n"
-                    "Please report this at https://github.com/pymmcore-plus/"
-                    "pymmcore-plus/issues/new",
-                    fg="bright_red",
-                )
-                raise typer.Exit(code=1) from None
-            install_path = dest / src.name
-            shutil.copytree(src, install_path, dirs_exist_ok=True)
-        finally:
-            subprocess.run(
-                ["hdiutil", "detach", mount], check=True, capture_output=True
-            )
-
-    # fix gatekeeper ... requires password
-    typer.secho(
-        "(Your password may be required to install Micro-manager.)",
-        fg=typer.colors.GREEN,
-    )
-    cmd = ["sudo", "xattr", "-r", "-d", "com.apple.quarantine", str(install_path)]
-    subprocess.run(cmd, check=True)
-
-    # # fix path randomization by temporarily copying elsewhere and back
-    with tempfile.TemporaryDirectory() as tmpdir:
-        _tmp = Path(tmpdir)
-        os.rename(install_path / "ImageJ.app", _tmp / "ImageJ.app")
-        os.rename(_tmp / "ImageJ.app", install_path / "ImageJ.app")
-
-
-def _available_versions() -> Dict[str, str]:
-    """Return a map of version -> url available for download."""
-    plat = {"Darwin": "Mac", "Windows": "Windows"}[PLATFORM]
-    with urlopen(f"{BASE_URL}/nightly/2.0/{plat}/") as resp:
-        html = resp.read().decode("utf-8")
-
-    return {
-        ref.rsplit("-", 1)[-1].split(".")[0]: BASE_URL + ref
-        for ref in re.findall(r"href=\"([^\"]+)\"", html)
-        if ref != "/"
-    }
-
-
-def _download_url(url: str, output_path: Path) -> None:
-    """Download `url` to `output_path` with a nice progress bar."""
-    import ssl
-
-    ctx = ssl.create_default_context()
-    ctx.check_hostname = False
-    ctx.verify_mode = ssl.CERT_NONE
-
-    pbar = progress.Progress(
-        progress.BarColumn(bar_width=None),
-        "[progress.percentage]{task.percentage:>3.1f}%",
-        "•",
-        progress.DownloadColumn(),
-    )
-    task_id = pbar.add_task("Download..", filename=output_path.name, start=False)
-
-    def hook(count: float, block_size: float, total_size: float) -> None:
-        pbar.update(task_id, total=int(total_size))
-        pbar.start_task(task_id)
-        pbar.update(task_id, advance=block_size)
-
-    print(f"[bold blue]Downloading {url} ...")
-    with pbar:
-        urlretrieve(url=url, filename=output_path, reporthook=hook)
+    pymmcore_plus.install._install(dest, release)
 
 
 def main() -> None:
